fix: use native i64 for timestamps

+++ programs/sponsored-cctp-src-periphery/src/event.rs
@@ -17,7 +17,7 @@ pub struct SponsoredDepositForBurn {
     pub quote_nonce: Vec<u8>, // Nonce is bytes32 random value, but it is more readable in logs expressed as encoded data blob.
     pub origin_sender: Pubkey,
     pub final_recipient: Pubkey,
-    pub quote_deadline: u32,
+    pub quote_deadline: i64,
     pub max_bps_to_sponsor: u64,
     pub final_token: Pubkey,
     pub signature: Vec<u8>, // This is fixed length, but using Vec so it is shown as encoded data blob in explorers.
